[🥦] deploy.sh pushed this!

+++ setup-remove.sh
@@ -3,14 +3,17 @@
 read -p "Sicher? Deine Bib & Datenbank wird gelöscht! (y/n): " answer
 
 if [[ "$answer" == "y" || "$answer" == "Y" ]]; then
-  docker compose -f compose.deploy.yml down --volumes
+	DIR="$HOME/.local/share/musikbruder"
 
-  rm -rf vendor
-  rm -rf node_modules
-  rm -rf public/data
-  rm -rf storage
-  rm -rf .env
-  rm -rf sql/last_migration
+	if [ ! -d $DIR ]; then
+		echo -e "musikbruder/ existiert nicht."
+		exit 1
+	fi
+
+	cd $DIR
+	docker compose -f compose.deploy.yml down --volumes
+	cd ..
+	rm -rf musikbruder
 else
-    echo "Abgebrochen :)"
+	echo "Abgebrochen :)"
 fi
